refactor(cli): renamed multiphase-mtx to three-phase
All the command inside multiphasemtx was starting with a three-phase already. I removed that and renamed the module to three-phase. I also nested the three-phase command under the multi-phase namespace to provide a better organization for the commands.
