+++ honeybee_radiance/cli/__init__.py
@@ -21,7 +21,6 @@ from .viewfactor import view_factor
 from .postprocess import post_process
 from .mtx import mtxop
 from .multiphase import multi_phase
-from .multiphasemtx import mpmtxop
 
 
 _logger = logging.getLogger(__name__)
@@ -73,7 +72,6 @@ radiance.add_command(view_factor)
 radiance.add_command(post_process, name='post-process')
 radiance.add_command(mtxop)
 radiance.add_command(multi_phase)
-radiance.add_command(mpmtxop)
 
 # add radiance sub-commands to honeybee CLI
 main.add_command(radiance)
